Isolated PhotoAuthor & added transitions on hover, updated logic for fullscreen photo rendering in Photo
