+++ src/components/photoComps/FullscreenPhoto/FullscreenPhoto.scss
@@ -1,6 +1,6 @@
 @use "../../../util/helpers/sassHelpers.scss" as *;
 
-#fullscreen-photo-container {
+.fullscreen-photo-container {
     @include width-height(100vw, 100vh);
     display: flex;
     justify-content: center;
