+++ src/components/photoComps/Photo/Photo.scss
@@ -1,32 +1,5 @@
 @use "../../../util/helpers/sassHelpers.scss" as *;
 
-.frame-container {
-
-    .author {
-        display: flex;
-        justify-content: space-between;
-        align-items: baseline;
-    }
-
-    .author a {
-        display: flex;
-        align-items: center;
-        column-gap: 10px;
-        text-align: left;
-        max-width: 80%;
-        font-size: 16px;
-        padding-top: 5px;
-        color: #4e2a2a;
-
-        img {
-            border-radius: 50%;
-        }
-
-        h2 {
-            font-weight: 400;
-        }
-    }
-
     .photograph {
         @include width-height(100%, auto);
         max-height: 80vh;
@@ -39,27 +12,11 @@
         cursor: pointer;
     }
 
-    .frame {
-
-        &::before {
-            content: "";
+    .photograph + address + button {
             position: absolute;
-            pointer-events: none;
-            inset: -4px;
-            border-radius: 2px;
-            box-shadow: 0 0.2vw 0.5vw 0.2vw rgba(0, 0, 0, 0.25) inset;
+            bottom: 10px;
+            right: 15px;
         }
-
-        &::after {
-            content: "";
-            position: absolute;
-            pointer-events: none;
-            inset: -7px;
-            border-radius: 2px;
-            box-shadow: 0 0.2vw 0.5vw 0.3vw rgba(0, 0, 0, 0.25);
-        }
-    }
-}
 
 .disable-click {
     pointer-events: none;
