+++ src/components/photoComps/PhotoAuthor/PhotoAuthor.scss
@@ -1,0 +1,38 @@
+.author a {
+    display: flex;
+    align-items: center;
+    column-gap: 10px;
+    text-align: left;
+    max-width: 80%;
+    font-size: 16px;
+    padding-top: 5px;
+    color: #4e2a2a;
+    cursor: default;
+
+    img {
+        border-radius: 50%;
+        cursor: pointer;
+    }
+
+    h2 {
+        font-weight: 400;
+        cursor: pointer;
+        position: relative;
+
+        &::after {
+            content: "";
+            position: absolute;
+            bottom: -4px;
+            left: 0;
+            background-color: #A99C9C;
+            width: 0;
+            height: 1px;
+            transition: width 0.3s ease-in-out;
+        }
+    }
+}
+
+.author a h2:hover::after {
+    width: 100%;
+    transition: width 0.3s ease-in-out;
+}
